EDIT: [Script] Mark Meshes last time used

+++ src/ResourceManager/ResourceTracker.cpp
@@ -7,19 +7,19 @@
 namespace RNGOEngine::Resources
 {
     void ResourceTracker::MarkModelLastUsed(
-        const Containers::Vectors::GenerationalKey<MeshResource>& modelKey, const int frame)
+        const Containers::Vectors::GenerationalKey<MeshResource>& modelKey, const size_t frame)
     {
         m_models[modelKey] = frame;
     }
 
     void ResourceTracker::MarkShaderProgramLastUsed(
-        const Containers::Vectors::GenerationalKey<Core::Renderer::ShaderProgramID>& shaderKey, const int frame)
+        const Containers::Vectors::GenerationalKey<Core::Renderer::ShaderProgramID>& shaderKey, const size_t frame)
     {
         m_shaderPrograms[shaderKey] = frame;
     }
 
     void ResourceTracker::MarkTextureLastUsed(
-        const Containers::Vectors::GenerationalKey<Core::Renderer::TextureID>& textureKey, const int frame)
+        const Containers::Vectors::GenerationalKey<Core::Renderer::TextureID>& textureKey, const size_t frame)
     {
         m_textures[textureKey] = frame;
     }
